refactor: organize book dtos into request and response
