+++ src/main/java/org/poolc/api/book/dto/request/CreateBookRequest.java
@@ -1,0 +1,4 @@
+package org.poolc.api.book.dto.request;
+
+public class CreateBookRequest {
+}
