+++ src/main/java/org/poolc/api/book/dto/request/UpdateBookRequest.java
@@ -1,0 +1,4 @@
+package org.poolc.api.book.dto.request;
+
+public class UpdateBookRequest {
+}
