+++ src/main/java/org/poolc/api/book/dto/response/BookResponse.java
@@ -1,0 +1,4 @@
+package org.poolc.api.book.dto;
+
+public class BookResponse {
+}
