+++ src/main/java/org/poolc/api/book/dto/response/NaverApiResponse.java
@@ -1,4 +1,4 @@
-package org.poolc.api.book.dto;
+package org.poolc.api.book.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
