+++ src/main/java/org/poolc/api/book/service/BookService.java
@@ -1,0 +1,4 @@
+package org.poolc.api.book.service;
+
+public class BookService {
+}
