+++ src/main/java/org/poolc/api/book/service/BookServiceImpl.java
@@ -1,0 +1,4 @@
+package org.poolc.api.book.service;
+
+public class BookServiceImpl {
+}
